backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
mysql-connector-python==9.0.0
Werkzeug==3.0.3
Pillow==10.4.0
PyMuPDF==1.24.9
pytesseract==0.3.10
